Split tests into disjoint sets
stratis_cert contains tests that are intended to test the CLI.
stratisd_cert contains tests that are intended to test the stratisd daemon.
